Refactor filters into their own directory
